requirements_dev.txt
black
black[jupyter]
build
bump-my-version
codespell
wheel
flake8
twine
click
pytest
pytest-runner

ipykernel
livereload
nbconvert
nbformat
sphinx
watchdog
mkdocs
mkdocs-git-revision-date-plugin 
mkdocs-git-revision-date-localized-plugin
mkdocs-jupyter>=0.24.0
mkdocs-material>=9.1.3 
mkdocs-pdf-export-plugin
mkdocstrings 
mkdocstrings-crystal
mkdocstrings-python-legacy
pygments
pymdown-extensions
numpyro
harmonic
pocomc


